--- incidents/txt_file_1
@@ -1,5 +1,0 @@
-                        Log File 1
-
-Please work.
-I beg of you!
-Yes!!
